Remove the old company business details page
