+++ src/apps/companies/apps/business-details/__test__/router.test.js
@@ -4,7 +4,7 @@ describe('Business details routes', () => {
   it('should define all routes', () => {
     const paths = router.stack.filter(r => r.route).map(r => r.route.path)
     expect(paths).to.deep.equal([
-      '/:companyId/business-details2',
+      '/:companyId/business-details',
     ])
   })
 })
