+++ src/apps/companies/apps/business-details/router.js
@@ -3,6 +3,6 @@ const router = require('express').Router()
 const urls = require('../../../../lib/urls')
 const { renderBusinessDetails } = require('./controllers')
 
-router.get(urls.companies.businessDetails2.route, renderBusinessDetails)
+router.get(urls.companies.businessDetails.route, renderBusinessDetails)
 
 module.exports = router
