+++ src/apps/companies/transformers/index.js
@@ -1,18 +1,12 @@
-const transformCompanyToBusinessHierarchyView = require('./company-to-business-hierarchy-view')
 const transformCompanyToExportDetailsView = require('./company-to-export-details-view')
-const transformCompanyToAboutView = require('./company-to-about-view')
 const transformCompanyToListItem = require('./company-to-list-item')
-const transformCompanyToOneListView = require('./company-to-one-list-view')
 const transformCompanyToSubsidiaryListItem = require('./company-to-subsidiary-list-item')
 const transformCoreTeamToCollection = require('./one-list-core-team-to-collection')
 const transformAccountManager = require('./company-to-account-manager')
 
 module.exports = {
-  transformCompanyToBusinessHierarchyView,
   transformCompanyToExportDetailsView,
-  transformCompanyToAboutView,
   transformCompanyToListItem,
-  transformCompanyToOneListView,
   transformCompanyToSubsidiaryListItem,
   transformCoreTeamToCollection,
   transformAccountManager,
