+++ test/selectors/company/subsidiaries.js
@@ -2,7 +2,7 @@ module.exports = () => {
   return {
     whyArchived: '[data-auto-id="subsidiariesWhyArchived"]',
     linkASubsidiary: '[data-auto-id="Link a subsidiary"]',
-    edit: '[data-auto-id="businessHierarchyDetailsContainer"] > a',
+    edit: '[data-auto-id="businessHierarchyDetailsContainer"] a:contains("Edit")',
     oneLinkedSubsidiary: 'a:contains("1 subsidiary")',
     linkASubsidiaryToHierarchy: 'a:contains("Link a subsidiary")',
     whyNoSubLink: 'span:contains("Why can I not link a subsidiary?")',
